+++ components/catalogs/back/src/ServiceAPI/Messenger/CommandBus.php
@@ -13,7 +13,7 @@ use Symfony\Component\Messenger\MessageBusInterface;
  * @copyright 2022 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-final class CommandBus
+final class CommandBus implements CommandBusInterface
 {
     use HandleTrait;