+++ components/catalogs/front/src/components/CatalogEdit/models/CatalogFormErrors.d.ts
@@ -1,0 +1,4 @@
+export type CatalogFormErrors = {
+    propertyPath: string;
+    message: string;
+}[];