+++ components/catalogs/front/src/components/ProductSelection/models/Family.d.ts
@@ -1,0 +1,4 @@
+export type Family = {
+    label: string;
+    code: string;
+};