+++ src/Akeneo/Tool/Bundle/ApiBundle/tests/integration/Controller/RootEndpointIntegration.php
@@ -2,7 +2,6 @@
 
 namespace Akeneo\Tool\Bundle\ApiBundle\tests\integration\Controller;
 
-use Akeneo\Test\Integration\Configuration;
 use Akeneo\Tool\Bundle\ApiBundle\tests\integration\ApiTestCase;
 use Symfony\Component\HttpFoundation\Response;